Refactor folder structure
Changing the folder structure to accomodate other folders that are not
necessarily source files. Moved all the source files into its own folder
under the name of the project thus following pythonic convention.
