+++ efficient/display.py
