+++ efficient/journald_logging.py
